Split WorkspacePath logic with Startup logic
